requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.2
django-badgify==0.4.0
django-bootstrap3==14.2.0
django-cleanup==5.1.0
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==3.1.0
djangorestframework==3.12.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
psycopg2==2.8.6
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0